+++ CHANGELOG.md
@@ -21,12 +21,13 @@ Security in case of vulnerabilities.
 ---
 
 ## Unreleased  
+
+### Added
+- added view significant date history `/projects/{projectId}/date-history`
 - added change significant date page "/projects/{projectId}/date-history/new"
 - added significant date change reason page "/projects/{projectId}/date-history/reason"
 - added confirm significant date change "/projects/{projectId}/date-history/reasons/later"
 
-### Added
-- New route `/projects/{projectId}/date-history`
 
 ### Changed